delete unused NetworkConfigDAO.kt
