+++ network/src/commonTest/kotlin/com/wire/kalium/api/ApiTest.kt
@@ -165,6 +165,6 @@ interface ApiTest {
     fun HttpRequestData.assertHttps() = assertEquals(expected = URLProtocol.HTTPS, actual = this.url.protocol)
 
     private companion object {
-        val TEST_BACKEND_CONFIG = BackendConfig("", "", "",  "",  "", "")
+        val TEST_BACKEND_CONFIG = BackendConfig("test.api.com", "test.account.com", "test.ws.com",  "test.blacklist",  "test.teams.com", "test.wire.com")
     }
 }
